Fold action_lib/utils.py into action_lib/__init__.py
Everything was directly imported, not a lot of code, nothing else used it
